refactor(ALL): IWYU everything
Include-what-you-use is a tool by google, which (based on clang) is able to generate for each source file a list of header files it should include. Thats based on which symbols (functions, vars..) are needed in that source file.
This is awesome!

Some conflicts:
- it put in a lot of boost/detail stuff -> manually removed that
- it put in includes for gtest internal stuff -> should have removed that (didnt yet)
- it put spdlog include in each sourcefile consuming Logging.hpp -> removed that
- it rearranged a lot of ICOMMANDs, which sometimes broke stuff
sometimes it forgot an include, but not many